Import bash configuration files.
Of note here is that the bash_path inventory variable is used to
populate the PATH on a per-host basis, rather than my old method that
used to determine the PATH at runtime based on a pre-configured list.

Along with a few other cleanups, this halves the time taken to parse my
shell initialisation, at least on my slow Octeon router (from 0.8
seconds down to 0.4).