Enable highlight search with customisation.
For many years I hated this option, and especially hated defaults that
enabled it (hi RedHat).  It's still awful with the default colour
scheme, especially when paired with the behaviour that you have to type
:noh to turn it back off (or search for some random string), but after
fixing the colour scheme and adding an Esc-Esc map to turn off the match
it works pretty well.